make_release_files.sh: Customize for common proprietary directory
This changes the script for the following features.
- make proprietary/socionext directory
- copy *.so to proprietary/socionext/{lib,lib64}/
- make symlinks to *.so for 'akebi96'
- copy OpenMAX heraders to proprietary/socionext/include/
- make Android.mk
- Add license and copyright descriptions

+++ make_release_files.sh
@@ -1,12 +1,16 @@
 #!/bin/bash -e
+# SPDX-License-Identifier: Apache-2.0
+# make_release_files.sh
+# Copyright (C) 2019 Socionext Inc.
 
-OUT_DIR=$OUT
-TARGET_DIR=proprietary
+VENDOR=socionext
+TARGET_DIR=proprietary/${VENDOR}
+TARGET_BOARD_NAME=sc1401aj1
 
 TO_EXTRACT_MULTI_LIB=(
-	"libomxprox.sc1401aj1.so \$(LOCAL_PATH)/include/OpenMAX/libomxil-prox/include"
-	"libsc.sc1401aj1.so \$(LOCAL_PATH)/include/libvideo-out/sc \$(LOCAL_PATH)/include/libvideo-out/sc/include \$(LOCAL_PATH)/include/libvideo-out/sc/include/reg"
-	"libvideo-out.sc1401aj1.so \$(LOCAL_PATH)/include/libvideo-out/voc_vout/include"
+	"libomxprox.\${TARGET_BOARD_NAME}.so OpenMAX/libomxil-prox/include"
+	"libsc.\${TARGET_BOARD_NAME}.so"
+	"libvideo-out.\${TARGET_BOARD_NAME}.so"
 )
 
 # --------------------------------------
@@ -14,15 +18,20 @@ TO_EXTRACT_MULTI_LIB=(
 # --------------------------------------
 Extract_files() {
 	echo \ \ Extracting files from Build Env
-	mkdir -p $TARGET_DIR/vendor/lib
-	mkdir -p $TARGET_DIR/vendor/lib64
+	mkdir -p ${TARGET_DIR}/lib
+	mkdir -p ${TARGET_DIR}/lib64
 
 	for ONE_FILE in "${TO_EXTRACT_MULTI_LIB[@]}"
 	do
-	 	data=(${ONE_FILE[@]})
-		echo copy $OUT_DIR/vendor/lib/${data[0]}
-		cp -a $OUT_DIR/vendor/lib/${data[0]}   $TARGET_DIR/vendor/lib
-		cp -a $OUT_DIR/vendor/lib64/${data[0]} $TARGET_DIR/vendor/lib64
+		data=(${ONE_FILE[@]})
+		FILE_NAME=$(eval echo ${data[0]})
+		echo copy ${OUT}/vendor/lib/${FILE_NAME}
+		cp -a ${OUT}/vendor/lib/${FILE_NAME}   ${TARGET_DIR}/lib
+		cp -a ${OUT}/vendor/lib64/${FILE_NAME} ${TARGET_DIR}/lib64
+		if [ x"${TARGET_BOARD}" != x"akebi96" ]; then
+			ln -sf ${FILE_NAME} ${TARGET_DIR}/lib/${FILE_NAME/${TARGET_BOARD_NAME}/akebi96}
+			ln -sf ${FILE_NAME} ${TARGET_DIR}/lib64/${FILE_NAME/${TARGET_BOARD_NAME}/akebi96}
+		fi
 	done
 }
 
@@ -30,64 +39,70 @@ Extract_files() {
 Make_Android_mk() {
 	echo \ \ make Android.mk
 
-	rm -f $TARGET_DIR/Android.mk
+	rm -f ${TARGET_DIR}/Android.mk
 
-	echo "LOCAL_PATH := \$(call my-dir)" >> $TARGET_DIR/Android.mk
-	echo "" >> $TARGET_DIR/Android.mk
+	echo "LOCAL_PATH := \$(call my-dir)" >> ${TARGET_DIR}/Android.mk
+	echo "" >> ${TARGET_DIR}/Android.mk
 	regexp="^(.*\.so?)[\s ](.*)"
 	for ONE_FILE in "${TO_EXTRACT_MULTI_LIB[@]}"
 	do
-	    if [[ ${ONE_FILE[@]} =~ $regexp ]] ; then
+		if [[ ${ONE_FILE[@]} =~ $regexp ]] ; then
 			FILE_NAME=${BASH_REMATCH[1]}
-			EXPORT_INCLUDE_NAME=${BASH_REMATCH[2]}
+			EXPORT_INCLUDE_NAME='$(LOCAL_PATH)/include/'${BASH_REMATCH[2]}
 		else
 			FILE_NAME=${ONE_FILE}
 			EXPORT_INCLUDE_NAME=""
 		fi
 		base_name=`basename -s .so ${FILE_NAME}`
-		echo "include \$(CLEAR_VARS)" >> $TARGET_DIR/Android.mk
-		echo "LOCAL_MULTILIB      := both" >> $TARGET_DIR/Android.mk
-		echo "LOCAL_MODULE        := ${base_name}" >> $TARGET_DIR/Android.mk
-		echo "LOCAL_SRC_FILES_32  := vendor/lib/${FILE_NAME}" >> $TARGET_DIR/Android.mk
-		echo "LOCAL_SRC_FILES_64  := vendor/lib64/${FILE_NAME}" >> $TARGET_DIR/Android.mk
-		echo "LOCAL_MODULE_CLASS  := SHARED_LIBRARIES" >> $TARGET_DIR/Android.mk
-		echo "LOCAL_MODULE_TAGS   := optional" >> $TARGET_DIR/Android.mk
-		echo "LOCAL_MODULE_SUFFIX := .so" >> $TARGET_DIR/Android.mk
+		echo "include \$(CLEAR_VARS)" >> ${TARGET_DIR}/Android.mk
+		echo "LOCAL_MULTILIB      := both" >> ${TARGET_DIR}/Android.mk
+		echo "LOCAL_MODULE        := ${base_name}" >> ${TARGET_DIR}/Android.mk
+		echo "LOCAL_SRC_FILES_32  := lib/${FILE_NAME}" >> ${TARGET_DIR}/Android.mk
+		echo "LOCAL_SRC_FILES_64  := lib64/${FILE_NAME}" >> ${TARGET_DIR}/Android.mk
+		echo "LOCAL_MODULE_CLASS  := SHARED_LIBRARIES" >> ${TARGET_DIR}/Android.mk
+		echo "LOCAL_MODULE_TAGS   := optional" >> ${TARGET_DIR}/Android.mk
+		echo "LOCAL_MODULE_SUFFIX := .so" >> ${TARGET_DIR}/Android.mk
 		if [ ! -z "${EXPORT_INCLUDE_NAME}" ] ; then
-			echo "LOCAL_EXPORT_C_INCLUDE_DIRS := ${EXPORT_INCLUDE_NAME}" >> $TARGET_DIR/Android.mk
+			echo "LOCAL_EXPORT_C_INCLUDE_DIRS := ${EXPORT_INCLUDE_NAME}" >> ${TARGET_DIR}/Android.mk
 		fi
-		echo "ifeq (\$(BOARD_VNDK_VERSION),current)" >> $TARGET_DIR/Android.mk
-		echo "LOCAL_VENDOR_MODULE := true" >> $TARGET_DIR/Android.mk
-		echo "endif" >> $TARGET_DIR/Android.mk
-		echo "include \$(BUILD_PREBUILT)" >> $TARGET_DIR/Android.mk
-		echo "" >> $TARGET_DIR/Android.mk
+		echo "ifeq (\$(BOARD_VNDK_VERSION),current)" >> ${TARGET_DIR}/Android.mk
+		echo "LOCAL_VENDOR_MODULE := true" >> ${TARGET_DIR}/Android.mk
+		echo "endif" >> ${TARGET_DIR}/Android.mk
+		echo "include \$(BUILD_PREBUILT)" >> ${TARGET_DIR}/Android.mk
+		echo "" >> ${TARGET_DIR}/Android.mk
 	done
 }
 
 copy_vendor_header_file() {
-	rsync -avm --delete \
-			--exclude="middle" \
-			--exclude="OpenMAX/libomxil-bellagio" \
-			--exclude=".git" \
-			--include='*.h' \
-		 	--include='*/' \
-			--exclude='*' \
-			$ANDROID_BUILD_TOP/vendor/socionext/sc1401aj1/* \
-			$TARGET_DIR/include
+	for ONE_FILE in "${TO_EXTRACT_MULTI_LIB[@]}"
+	do
+		data=(${ONE_FILE[@]})
+		EXPORT_INCLUDE_NAME=${data[1]}
+		if [ x"${EXPORT_INCLUDE_NAME}" != x"" ]; then
+			mkdir -p ${TARGET_DIR}/include/${EXPORT_INCLUDE_NAME}
+			rsync -avm --delete \
+			      --exclude='.git' \
+			      --include='*.h' \
+			      --include='*/' \
+			      --exclude='*' \
+			      ${ANDROID_BUILD_TOP}/vendor/${VENDOR}/${TARGET_BOARD_NAME}/${EXPORT_INCLUDE_NAME}/* \
+			      ${TARGET_DIR}/include/${EXPORT_INCLUDE_NAME}
+		fi
+	done
 }
 
 # ------------------------
 # main
 # ------------------------
-if [ -z $OUT ]; then
+if [ -z ${OUT} ]; then
 	echo "Please select a build valiant using lunch command"
 	exit 1;
 fi
 
-if [ -e $TARGET_DIR ] ; then
-	rm -rf $TARGET_DIR
+if [ -e ${TARGET_DIR} ] ; then
+	rm -rf ${TARGET_DIR}
 fi
-mkdir -p $TARGET_DIR
+mkdir -p ${TARGET_DIR}
 
 copy_vendor_header_file
 Extract_files
